# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: EDP Control Plane API
  description: |
    Enterprise data control plane for registering sources and datasets,
    authoring and deploying pipelines, enforcing data quality and governance
    policy, tracking lineage and impact, orchestrating runs, and recording an
    append-only audit trail.

    ## Conventions

    * All collection endpoints are paginated with `limit` and `offset`, and
      sorted with `sort_by` and `sort_dir`. See the shared parameters.
    * Errors always use the same envelope: `error.code`, `error.message`,
      `error.details` and a top-level `request_id` that matches the
      `X-Request-Id` response header.
    * Mutations are `PATCH` rather than `PUT`, so a caller updates only the
      fields it means to change.
    * Request bodies are validated before any state is written; a validation
      failure returns `422` with one entry per offending field.
  version: 0.1.0
  contact:
    name: Data Platform Engineering
servers:
  - url: http://localhost:8080/api/v1
    description: Local development
  - url: https://edp.internal/api/v1
    description: Internal deployment
security:
  - bearerAuth: []
...
paths:
...
  /datasets:
    get:
      tags: [datasets]
      operationId: listDatasets
      summary: List datasets
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/SortBy'
        - $ref: '#/components/parameters/SortDir'
        - name: q
          in: query
          schema: { type: string }
        - name: kind
          in: query
          schema: { type: string, enum: [table, view, file, topic, stream, model] }
        - name: classification
          in: query
          schema: { type: string, enum: [public, internal, confidential, restricted, pii] }
        - name: status
          in: query
          schema: { type: string }
        - name: source_id
          in: query
          schema: { type: string, format: uuid }
        - name: domain
          in: query
          schema: { type: string }
        - name: owner_id
          in: query
          schema: { type: string }
        - name: steward_id
          in: query
          schema: { type: string }
        - name: include_columns
          in: query
          description: Include the column list on each dataset.
          schema: { type: boolean, default: false }
        - name: tag
          in: query
          schema: { type: array, items: { type: string } }
      responses:
        '200':
          description: A page of datasets.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DatasetPage' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [datasets]
      operationId: createDataset
      summary: Register a dataset
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DatasetWrite' }
      responses:
        '201':
          description: The registered dataset.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Dataset' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
  /datasets/by-slug/{slug}:
    get:
      tags: [datasets]
      operationId: getDatasetBySlug
      summary: Fetch a dataset by slug
      parameters:
        - $ref: '#/components/parameters/Slug'
        - name: include_columns
          in: query
          schema: { type: boolean, default: false }
      responses:
        '200':
          description: The dataset.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Dataset' }
        '404': { $ref: '#/components/responses/NotFound' }
  /datasets/{datasetID}:
    parameters:
      - $ref: '#/components/parameters/DatasetID'
    get:
      tags: [datasets]
      operationId: getDataset
      summary: Fetch a dataset
      parameters:
        - name: include_columns
          in: query
          schema: { type: boolean, default: false }
      responses:
        '200':
          description: The dataset.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Dataset' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [datasets]
      operationId: updateDataset
      summary: Update a dataset
      description: |
        Destructive changes are refused. Classification may only be raised, never
        lowered, and columns cannot be dropped through this endpoint.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DatasetWritePartial' }
      responses:
        '200':
          description: The updated dataset.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Dataset' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
    delete:
      tags: [datasets]
      operationId: deleteDataset
      summary: Delete a dataset
      responses:
        '204': { description: Deleted. }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
...
  /audit-events:
    get:
      tags: [audit]
      operationId: listAuditEvents
      summary: List audit events
      description: |
        Every query is bounded. `since` is required in effect: when omitted the
        service applies its maximum lookback window rather than scanning the whole
        table.
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/SortBy'
        - $ref: '#/components/parameters/SortDir'
        - name: entity_type
          in: query
          schema: { type: string }
        - name: entity_id
          in: query
          schema: { type: string, format: uuid }
        - name: action
          in: query
          schema: { type: string }
        - name: actor_id
          in: query
          schema: { type: string }
        - name: outcome
          in: query
          schema: { type: string, enum: [success, failure, denied] }
        - name: since
          in: query
          schema: { type: string, format: date-time }
        - name: until
          in: query
          schema: { type: string, format: date-time }
      responses:
        '200':
          description: A page of audit events.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuditEventPage' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /audit-events/{eventID}:
    parameters:
      - name: eventID
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [audit]
      operationId: getAuditEvent
      summary: Fetch an audit event
      responses:
        '200':
          description: The event.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuditEvent' }
        '404': { $ref: '#/components/responses/NotFound' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        A JWT verified against the configured JWKS. Issuer, audience and expiry are
        all checked. When authentication is disabled the API runs without any
        middleware and every request is attributed to a development actor.
  parameters:
    Limit:
      name: limit
      in: query
      description: Page size.
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    Offset:
      name: offset
      in: query
      description: Rows to skip.
      schema: { type: integer, minimum: 0, default: 0 }
    SortBy:
      name: sort_by
      in: query
      description: Column to sort on. Unknown columns are rejected.
      schema: { type: string }
    SortDir:
      name: sort_dir
      in: query
      schema: { type: string, enum: [asc, desc], default: desc }
    Slug:
      name: slug
      in: path
      required: true
      schema: { type: string, maxLength: 100 }
...
    DatasetID:
      name: datasetID
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    BadRequest:
      description: The query string or path parameters were malformed.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: No valid token was presented.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: No such resource.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflict:
      description: The request conflicts with current state.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    ValidationFailed:
      description: The body failed validation. Every offending field is listed.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      required: [error, request_id]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Stable machine-readable code.
              enum:
                - BAD_REQUEST
                - UNAUTHORIZED
                - FORBIDDEN
                - NOT_FOUND
                - METHOD_NOT_ALLOWED
                - CONFLICT
                - VALIDATION_FAILED
                - RATE_LIMITED
                - INTERNAL_ERROR
                - SERVICE_UNAVAILABLE
                - NOT_IMPLEMENTED
            message:
              type: string
              description: Human-readable summary. Safe to show a user.
            details:
              type: array
              description: Present for validation failures.
              items:
                type: object
                properties:
                  field: { type: string }
                  message: { type: string }
        request_id:
          type: string
          format: uuid
          description: Matches the X-Request-Id header, for correlating with logs.
    PageMeta:
      type: object
      required: [limit, offset, total]
      properties:
        limit: { type: integer }
        offset: { type: integer }
        total: { type: integer }
...
    Dataset:
      type: object
      properties:
        id: { type: string, format: uuid }
        source_id: { type: string, format: uuid }
        name: { type: string }
        slug: { type: string }
        description: { type: string }
        kind: { type: string }
        environment: { type: string }
        classification: { type: string }
        sensitivity: { type: string }
        status: { type: string }
        domain: { type: string }
        owner_id: { type: string }
        steward_id: { type: string }
        tags: { type: array, items: { type: string } }
        metadata: { type: object, additionalProperties: true }
        columns:
          type: array
          description: Populated only when include_columns is true.
          items: { $ref: '#/components/schemas/DatasetColumn' }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    DatasetColumn:
      type: object
      properties:
        name: { type: string }
        data_type: { type: string }
        nullable: { type: boolean }
        primary_key: { type: boolean }
        classification: { type: string }
        description: { type: string }
        ordinal: { type: integer }
    DatasetWrite:
      type: object
      required: [source_id, name, kind]
      properties:
        source_id: { type: string, format: uuid }
        name: { type: string }
        slug: { type: string }
        description: { type: string }
        kind: { type: string }
        environment: { type: string }
        classification: { type: string }
        sensitivity: { type: string }
        domain: { type: string }
        owner_id: { type: string }
        steward_id: { type: string }
        tags: { type: array, items: { type: string } }
        metadata: { type: object, additionalProperties: true }
        columns: { type: array, items: { $ref: '#/components/schemas/DatasetColumn' } }
    DatasetWritePartial:
      allOf:
        - $ref: '#/components/schemas/DatasetWrite'
...
    AuditEvent:
      type: object
      properties:
        id: { type: string, format: uuid }
        occurred_at: { type: string, format: date-time }
        actor_id: { type: string }
        actor_type: { type: string }
        action: { type: string }
        entity_type: { type: string }
        entity_id: { type: string, format: uuid, nullable: true }
        entity_name: { type: string }
        outcome: { type: string, enum: [success, failure, denied] }
        change_set:
          type: object
          additionalProperties: true
          description: Redacted and size-bounded before storage.
        duration_ms: { type: integer, nullable: true }
        ip_address: { type: string, nullable: true }
        user_agent: { type: string }
        request_id: { type: string }
...
    DatasetPage:
      allOf:
        - $ref: '#/components/schemas/Page'
        - type: object
          properties:
            items: { type: array, items: { $ref: '#/components/schemas/Dataset' } }
...
    AuditEventPage:
      allOf:
        - $ref: '#/components/schemas/Page'
        - type: object
          properties:
            items: { type: array, items: { $ref: '#/components/schemas/AuditEvent' } }
    Page:
      type: object
      description: Every list endpoint returns the same envelope.
      required: [items, meta]
      properties:
        items:
          type: array
          items: {}
        meta: { $ref: '#/components/schemas/PageMeta' }